requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
filelock==3.9.0
fillpdf==0.7.2
idna==3.4
numpy==1.24.2
packaging==23.0
pdf2image==1.17.0
pdfrw2==0.5.0
pillow==10.2.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
